google_vision/requirements.txt
google.cloud.vision
google.cloud.storage
google.cloud.firestore